Fixed bug, measuring QTL properties if none exist. Updated plotting script.

+++ ServerScripts/Polyself_Q_Sims.sh
@@ -3,13 +3,12 @@
 # 4th December 2019
 # Modification of Gertjan Bisschop's script for running array job on server
 # SLiM code of polygenic selection model
-# Need to first run 'source activate fscne-env'
 
 # 29th Jan 2020
-# Runs sims
+# Runs sims only
 
 # Grid Engine options (lines prefixed with #$)
-#$ -N Polysel_Self
+#$ -N Polysel_Self_Sims
 #$ -V
 #$ -cwd
 #$ -t 1-24 		# Run command for each line of parameter file
@@ -25,7 +24,12 @@ NTR=$(sed -n ${SGE_TASK_ID}p /data/hartfield/polyself/scripts/PolyselParameters.
 MVAR=$(sed -n ${SGE_TASK_ID}p /data/hartfield/polyself/scripts/PolyselParameters.txt | awk '{print $6}')
 
 # Running simulations, parameters in 'PolyselParameters.txt'
-rm -rf /scratch/mhartfield/polyself_out/
-mkdir /scratch/mhartfield/polyself_out/
-mkdir /scratch/mhartfield/polyself_out/data/
+if [ $SGE_TASK_ID == $SGE_TASK_FIRST ]
+then
+	rm -rf /scratch/mhartfield/polyself_out/
+	mkdir /scratch/mhartfield/polyself_out/
+	mkdir /scratch/mhartfield/polyself_out/data/
+else
+	sleep 10
+fi
 /ceph/software/slim/slim_v3.3/SLiM/slim -d s=$SEL -d h=$DOM -d sfrate=$SELF -d newo=$NEWOP -d nt=$NTR -d mvar=$MVAR /data/hartfield/polyself/scripts/Polygenic_Selection_With_Selfing.slim
